+++ backend/tests/api/test_attendance_record_create_view.py
@@ -3,7 +3,7 @@ from rest_framework.test import APIClient
 from rest_framework import status
 from datetime import date, time
 from apps.attendance.models.record import Record
-from tests.factory.attendance_record_factory import RecordFactory
+from tests.factory.record_factory import RecordFactory
 from tests.factory.work_pattern_factory import WorkPatternFactory
 from tests.factory.break_setting_factory import BreakSettingFactory
 from tests.factory.user_factory import UserFactory
