+++ backend/tests/unit/test_attendance_record_model.py
@@ -2,7 +2,7 @@ import pytest
 from datetime import date, time
 from apps.attendance.models.record import Record
 from apps.utility.enums import WorkStatus
-from tests.factory.attendance_record_factory import RecordFactory
+from tests.factory.record_factory import RecordFactory
 from tests.factory.user_factory import UserFactory
 from tests.factory.work_pattern_factory import WorkPatternFactory
